backend_old/pyproject.toml
[project]
name = "cassidy-gemini2.5-backend"
version = "0.1.0"
dependencies = [
    "typer",
    "anthropic==0.49.0",
    "fastapi==0.115.12",
    "pydantic==2.11.1",
    "pydantic-settings==2.8.1",
    "uvicorn==0.34.0",
    "pytest==8.3.5",
    "httpx==0.28.1",
    "starlette==0.46.1",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
    "anyio==4.9.0",
    "click==8.1.8",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jiter==0.9.0",
    "packaging==24.2",
    "pluggy==1.5.0",
    "sniffio==1.3.1",
    "typing-extensions==4.13.1",
    "typing-inspection==0.4.0",
    "uvloop==0.21.0",
    "certifi==2025.1.31",
    "distro==1.9.0",
    "annotated-types==0.7.0"
]

[tool.ruff]
line-length = 88
target-version = "py311" # Match tech-architecture rule

[tool.ruff.lint]
select = ["E", "F", "W", "I"] # Enable standard checks + isort
ignore = []

[tool.ruff.format]
quote-style = "double" 